Tue, Oct 3, 2023, 1:58 PM +01:00

+++ filter-less/helpers.c
@@ -129,8 +129,8 @@ void blur(int height, int width, RGBTRIPLE image[height][width])
             }
 
                 image[i][j].rgbtBlue = round((float)(sumBlue/count));
-                image[i][j].rgbtGreen = round((float)(sumGreen/count);
-                image[i][j].rgbtRed = round(sumRed/count);
+                image[i][j].rgbtGreen = round((float)(sumGreen/count));
+                image[i][j].rgbtRed = round((float)(sumRed/count));
         }
     }
